+++ com.microsoft.copilot.eclipse.core/src/com/microsoft/copilot/eclipse/core/lsp/protocol/LanguageModelToolResult.java
@@ -11,6 +11,11 @@ import org.eclipse.lsp4j.jsonrpc.util.ToStringBuilder;
  */
 public class LanguageModelToolResult {
   /**
+   * The status of the tool invocation.
+   */
+  private String status;
+
+  /**
    * A list of tool result content parts. Includes `unknown` because this list may be extended with new content types in
    * the future.
    */
@@ -20,13 +25,15 @@ public class LanguageModelToolResult {
    * Creates a new LanguageModelToolResult.
    */
   public LanguageModelToolResult() {
+    this.status = ToolInvocationStatus.success.toString();
     this.content = new ArrayList<>();
   }
 
   /**
-   * Creates a new LanguageModelToolResult with content.
+   * Creates a new LanguageModelToolResult with content and ToolInvocationStatus.
    */
-  public LanguageModelToolResult(String resultContent) {
+  public LanguageModelToolResult(String resultContent, ToolInvocationStatus status) {
+    this.status = status.toString();
     this.content = new ArrayList<>();
     this.content.add(new LanguageModelTextPart(resultContent));
   }
@@ -49,9 +56,17 @@ public class LanguageModelToolResult {
     this.content = content;
   }
 
+  public ToolInvocationStatus getStatus() {
+    return ToolInvocationStatus.valueOf(status);
+  }
+
+  public void setStatus(ToolInvocationStatus status) {
+    this.status = status.toString();
+  }
+
   @Override
   public int hashCode() {
-    return Objects.hash(content);
+    return Objects.hash(content, status);
   }
 
   @Override
@@ -66,13 +81,21 @@ public class LanguageModelToolResult {
       return false;
     }
     LanguageModelToolResult other = (LanguageModelToolResult) obj;
-    return Objects.equals(content, other.content);
+    return Objects.equals(content, other.content) && status == other.status;
   }
 
   @Override
   public String toString() {
     ToStringBuilder builder = new ToStringBuilder(this);
+    builder.add("status", status);
     builder.add("content", content);
     return builder.toString();
+  }
+
+  /**
+   * The status of the tool invocation.
+   */
+  public static enum ToolInvocationStatus {
+    success, error, cancelled
   }
 }